+++ docs/releases.md
@@ -33,8 +33,9 @@ Patches:
 * Fixing `start` parameter that is not supported for `getSchemas` anymore.
 * default the getSchemas to not get the adhoc schema.
 * Fixing the FieldGroupManager intantiation.
-* adding documentation on the Catalog
-* adding 
+* Adding documentation on the Catalog
+* Adding `onlyDestinations` and `onlySources` as parameter for the `getFlows` method.
+* Fixing the `createDataset` method when wanting to have a dataset enabled for Profile and Identity Service. 
 
 ## version 0.3.0